speedtest/requirements.txt
fluent-logger
speedtest-cli
